MANOVA and MANCOVA
Embark on a journey of statistical exploration with our repository, where we delve into the intricacies of student performance analysis using multivariate multiple regression and MANOVA/MANCOVA techniques. Through Python scripts and accompanying explanations, we unravel the complex relationships between academic variables, program types, and socioeconomic factors. Our mission is to provide educators, policymakers, and researchers with actionable insights to enhance educational outcomes and drive positive change in academic settings. Join us as we navigate the dynamic landscape of student performance analysis and empower stakeholders with the tools and knowledge needed to shape the future of education.